// src/app/services/data/model/manage-payment-methods.ts
export class CC {
    NickName: string;
    Default: boolean;
    FirstName: string;
    LastName: string;
    Address1: string;
    Address2: string;
    State: string;
    City: string;
    PostalCode: string;
    CountryCode: string;
    Number: string;
    ExpirationMonth: string;
    ExpirationYear: string;
    Cvv: string;
    expdate: string;
  }
  
  export class ACH {
    NickName: string;
    Default: boolean;
    FirstName: string;
    LastName: string;
    Address1: string;
    Address2: string;
    State: string;
    City: string;
    PostalCode: string;
    CountryCode: string;
    RoutingNumber: string;
    AccountNumber: string;
    AccountType: string;
  }
  